+++ src/main/java/pepjebs/mapatlases/mixin/MapRendererMixin.java
@@ -13,7 +13,7 @@ public class MapRendererMixin {
     @Redirect(method = "draw(Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;ZI)V",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/client/util/math/MatrixStack;scale(FFF)V"))
     private void scaleProxy(MatrixStack matrices, float x, float y, float z) {
-        int multiplier = MapAtlasesClient.getWorldMapZoomLevel();
+        float multiplier = MapAtlasesClient.getWorldMapZoomLevel();
         matrices.scale(x*multiplier, y*multiplier, z);
     }
 }
